Add viterbi module
