+++ lindera-core/src/dictionary.rs
@@ -2,8 +2,6 @@ pub mod character_definition;
 pub mod connection_cost_matrix;
 pub mod prefix_dictionary;
 pub mod unknown_dictionary;
-pub mod viterbi;
-pub mod word_entry;
 
 use std::str;
