+++ lindera-core/src/mode.rs
@@ -2,8 +2,8 @@ use std::str::FromStr;
 
 use serde::{Deserialize, Serialize};
 
-use crate::dictionary::viterbi::Edge;
 use crate::error::{LinderaError, LinderaErrorKind};
+use crate::viterbi::Edge;
 
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq)]
 pub struct Penalty {
