+++ lindera/src/token.rs
@@ -2,8 +2,8 @@ use std::borrow::Cow;
 
 use serde::Serialize;
 
-use lindera_core::dictionary::word_entry::WordId;
 use lindera_core::dictionary::{Dictionary, UserDictionary, UNK};
+use lindera_core::viterbi::WordId;
 
 #[derive(Serialize, Clone)]
 pub struct Token<'a> {
