+++ lindera/src/token_filter/japanese_base_form.rs
@@ -196,7 +196,7 @@ mod tests {
     fn test_japanese_base_form_token_filter_apply_ipadic() {
         use std::borrow::Cow;
 
-        use lindera_core::dictionary::word_entry::WordId;
+        use lindera_core::viterbi::WordId;
 
         use crate::dictionary::{load_dictionary_from_kind, DictionaryKind};
         use crate::token::Token;
@@ -313,7 +313,7 @@ mod tests {
     fn test_japanese_base_form_token_filter_apply_unidic() {
         use std::borrow::Cow;
 
-        use lindera_core::dictionary::word_entry::WordId;
+        use lindera_core::viterbi::WordId;
 
         use crate::dictionary::{load_dictionary_from_kind, DictionaryKind};
         use crate::token::Token;
