+++ lindera/src/token_filter/japanese_number.rs
@@ -935,7 +935,7 @@ mod tests {
     fn test_japanese_number_token_filter_apply_numbers_ipadic() {
         use std::borrow::Cow;
 
-        use lindera_core::dictionary::word_entry::WordId;
+        use lindera_core::viterbi::WordId;
 
         use crate::dictionary::{load_dictionary_from_kind, DictionaryKind};
         use crate::token_filter::japanese_number::JapaneseNumberTokenFilter;
@@ -1100,7 +1100,7 @@ mod tests {
     fn test_japanese_number_token_filter_apply_empty_ipadic() {
         use std::borrow::Cow;
 
-        use lindera_core::dictionary::word_entry::WordId;
+        use lindera_core::viterbi::WordId;
 
         use crate::dictionary::{load_dictionary_from_kind, DictionaryKind};
         use crate::token_filter::japanese_number::JapaneseNumberTokenFilter;
