+++ lindera/src/token_filter/japanese_stop_tags.rs
@@ -233,7 +233,7 @@ mod tests {
     fn test_japanese_stop_tags_token_filter_apply_ipadic() {
         use std::borrow::Cow;
 
-        use lindera_core::dictionary::word_entry::WordId;
+        use lindera_core::viterbi::WordId;
 
         use crate::dictionary::{load_dictionary_from_kind, DictionaryKind};
         use crate::token::Token;
