+++ lindera/src/token_filter/keep_words.rs
@@ -130,7 +130,7 @@ mod tests {
     fn test_keep_words_token_filter_apply_ipadic() {
         use std::borrow::Cow;
 
-        use lindera_core::dictionary::word_entry::WordId;
+        use lindera_core::viterbi::WordId;
 
         use crate::dictionary::{load_dictionary_from_kind, DictionaryKind};
         use crate::token::Token;
