+++ lindera/src/token_filter/korean_stop_tags.rs
@@ -176,7 +176,7 @@ mod tests {
     fn test_korean_stop_tags_token_filter_apply() {
         use std::borrow::Cow;
 
-        use lindera_core::dictionary::word_entry::WordId;
+        use lindera_core::viterbi::WordId;
 
         use crate::dictionary::{load_dictionary_from_kind, DictionaryKind};
         use crate::token::Token;
