+++ lindera/src/token_filter/lowercase.rs
@@ -45,7 +45,7 @@ mod tests {
     fn test_lowercase_token_filter_apply_ipadic() {
         use std::borrow::Cow;
 
-        use lindera_core::dictionary::word_entry::WordId;
+        use lindera_core::viterbi::WordId;
 
         use crate::dictionary::{load_dictionary_from_kind, DictionaryKind};
         use crate::token::Token;
