+++ lindera/src/token_filter/stop_words.rs
@@ -104,7 +104,7 @@ mod tests {
     fn test_stop_words_token_filter_apply_ipadic() {
         use std::borrow::Cow;
 
-        use lindera_core::dictionary::word_entry::WordId;
+        use lindera_core::viterbi::WordId;
 
         use crate::dictionary::{load_dictionary_from_kind, DictionaryKind};
         use crate::token::Token;
